# zip-repl/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GEF Player Statistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a new player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRequest"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    put:
      operationId: updatePlayer
      tags: [players]
      summary: Update player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerRequest"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /players/{id}/stats:
    get:
      operationId: getPlayerStats
      tags: [players]
      summary: Get detailed player stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        teamId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        marketValue:
          type: number
          nullable: true
        salary:
          type: number
          nullable: true
        matchesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        goalsScored:
          type: integer
        goalsConceded:
          type: integer
        mvpCount:
          type: integer
        overallRating:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - matchesPlayed
        - wins
        - losses
        - draws
        - goalsScored
        - goalsConceded
        - mvpCount
        - overallRating
        - createdAt
    CreatePlayerRequest:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        teamId:
          type: integer
          nullable: true
        nationality:
          type: string
          nullable: true
        marketValue:
          type: number
          nullable: true
        salary:
          type: number
          nullable: true
      required:
        - name
    UpdatePlayerRequest:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        teamId:
          type: integer
          nullable: true
        nationality:
          type: string
          nullable: true
        marketValue:
          type: number
          nullable: true
        salary:
          type: number
          nullable: true
    PlayerStats:
      type: object
      properties:
        playerId:
          type: integer
        name:
          type: string
        matchesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        winRate:
          type: number
        goalsScored:
          type: integer
        goalsConceded:
          type: integer
        goalDiff:
          type: integer
        goalsPerMatch:
          type: number
        goalsConcededPerMatch:
          type: number
        mvpCount:
          type: integer
        overallRating:
          type: number
        recentMatches:
          type: array
          items:
            $ref: "#/components/schemas/MatchSummary"
        awards:
          type: array
          items:
            $ref: "#/components/schemas/Award"
      required:
        - playerId
        - name
        - matchesPlayed
        - wins
        - losses
        - draws
        - winRate
        - goalsScored
        - goalsConceded
        - goalDiff
        - goalsPerMatch
        - goalsConcededPerMatch
        - mvpCount
        - overallRating
        - recentMatches
        - awards
    MatchSummary:
      type: object
      properties:
        matchId:
          type: integer
        date:
          type: string
        opponentName:
          type: string
        playerGoals:
          type: integer
        opponentGoals:
          type: integer
        result:
          type: string
      required:
        - matchId
        - date
        - opponentName
        - playerGoals
        - opponentGoals
        - result
...
    Award:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        awardedAt:
          type: string
      required:
        - id
        - playerId
        - title
        - awardedAt
...